Fix white screen: wrap ClientSideSuspense in a client component
ClientSideSuspense uses useState and must run inside a "use client"
boundary. Moving it directly into the server component page.tsx caused
a useState error during SSR. Extract it into GameContent.tsx ("use client")
which is imported by page.tsx instead.
